Update fbd.html
deleted some code :
<p align="center">The picture above shows that the string must provide the necessary centripetal force to move the ball in a circle. 
If the string breaks, the ball will move off in a straight line. The straight line motion in the absence 
of the constraining force is an example of Newton's first law. The example here presumes that no other net forces are 
acting, such as horizontal motion on a frictionless surface. The vertical circle is more involved.</p>